% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmyloginpath.R
\name{login_parse}
\alias{login_parse}
\title{Decrypt and parse login path file}
\usage{
login_parse(login_path, path = NULL)
}
\arguments{
\item{login_path}{name of server to connect to}

\item{path}{optional, path to login path file.}
}
\value{
List with login details
}
\description{
Decrypt and parse login path file
}
