% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmyloginpath.R
\name{login_read}
\alias{login_read}
\title{Decrypt login path file}
\usage{
login_read(path = NULL)
}
\arguments{
\item{path}{optional, path to login path file.}
}
\value{
String containing decrypted login paths in INI format
}
\description{
Decrypt login path file
}
