% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{read.mysql.ini}
\alias{read.mysql.ini}
\title{Read and parse .ini file to list}
\usage{
read.mysql.ini(fileText, encoding = getOption("encoding"))
}
\arguments{
\item{fileText}{text of file to parse}

\item{encoding}{Encoding of filepath parameter, will default to system
encoding if not specifield}
}
\value{
List with length equivalent to number of \link{sections}, each section is
a new list
}
\description{
Read and parse .ini file to list
}
\details{
Lines starting with '#' or ';' are comments and will not be parsed
}
